# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/profiles/{id}": {
      "get": {
        "description": "특정 프로필의 상세 정보를 조회합니다.",
        "operationId": "ProfileController_getProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프로필 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "프로필 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "홍길동"
                    },
                    "role": {
                      "type": "string",
                      "example": "STUDENT"
                    },
                    "birthDate": {
                      "type": "string",
                      "example": "1990-01-01"
                    },
                    "userId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "프로필을 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "프로필 상세 조회",
        "tags": [
          "profiles"
        ]
      }
    },
...
    "/organizations/{id}": {
      "get": {
        "description": "특정 조직의 상세 정보를 조회합니다.",
        "operationId": "OrganizationController_getOrganization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조직 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "테스트 학교"
                    },
                    "type": {
                      "type": "string",
                      "example": "school"
                    },
                    "mainAdminRole": {
                      "type": "string",
                      "example": "MAIN_ADMIN"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "조직을 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "조직 상세 조회",
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/permissions/organizations/{organizationId}": {
      "post": {
        "description": "조직 내 특정 프로필에 권한을 부여합니다.",
        "operationId": "PermissionController_grantPermission",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantPermissionPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "권한 부여 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "organizationId": {
                      "type": "number",
                      "example": 1
                    },
                    "profileId": {
                      "type": "number",
                      "example": 1
                    },
                    "permission": {
                      "type": "string",
                      "example": "CREATE_COURSE"
                    },
                    "grantedByProfileId": {
                      "type": "number",
                      "example": 1
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "expiresAt": {
                      "type": "string",
                      "example": "2024-12-31T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "조직 또는 프로필을 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "권한 부여",
        "tags": [
          "permissions"
        ]
      },
      "get": {
        "description": "조직의 모든 권한을 조회합니다.",
        "operationId": "PermissionController_getOrganizationPermissions",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조직 권한 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "organizationId": {
                        "type": "number",
                        "example": 1
                      },
                      "profileId": {
                        "type": "number",
                        "example": 1
                      },
                      "permission": {
                        "type": "string",
                        "example": "CREATE_COURSE"
                      },
                      "grantedByProfileId": {
                        "type": "number",
                        "example": 1
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-01-01T00:00:00.000Z"
                      },
                      "expiresAt": {
                        "type": "string",
                        "example": "2024-12-31T00:00:00.000Z"
                      },
                      "organization": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "테스트 학교"
                          }
                        }
                      },
                      "profile": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "홍길동"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "조직 권한 목록 조회",
        "tags": [
          "permissions"
        ]
      }
    },
    "/permissions/organizations/{organizationId}/profiles/{profileId}": {
      "delete": {
        "description": "조직 내 특정 프로필의 권한을 취소합니다.",
        "operationId": "PermissionController_revokePermission",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "프로필 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "permission",
            "required": true,
            "in": "query",
            "description": "취소할 권한",
            "schema": {
              "enum": [
                "course:create",
                "course:update",
                "course:delete",
                "course:enrollment:manage",
                "course:attendance:manage",
                "course:instructor:assign",
                "course:view",
                "course:session:manage",
                "course:class:manage",
                "permission:manage"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "권한 취소 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "권한이 성공적으로 취소되었습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "권한을 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "권한 취소",
        "tags": [
          "permissions"
        ]
      }
    },
    "/permissions/{id}": {
      "put": {
        "description": "기존 권한 정보를 수정합니다.",
        "operationId": "PermissionController_updatePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "권한 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "권한 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "organizationId": {
                      "type": "number",
                      "example": 1
                    },
                    "profileId": {
                      "type": "number",
                      "example": 1
                    },
                    "permission": {
                      "type": "string",
                      "example": "CREATE_COURSE"
                    },
                    "expiresAt": {
                      "type": "string",
                      "example": "2024-12-31T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "권한을 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "권한 수정",
        "tags": [
          "permissions"
        ]
      }
    },
    "/permissions/profiles/{profileId}": {
      "get": {
        "description": "특정 프로필의 모든 권한을 조회합니다.",
        "operationId": "PermissionController_getProfilePermissions",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "프로필 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "프로필 권한 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "organizationId": {
                        "type": "number",
                        "example": 1
                      },
                      "profileId": {
                        "type": "number",
                        "example": 1
                      },
                      "permission": {
                        "type": "string",
                        "example": "CREATE_COURSE"
                      },
                      "grantedByProfileId": {
                        "type": "number",
                        "example": 1
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-01-01T00:00:00.000Z"
                      },
                      "expiresAt": {
                        "type": "string",
                        "example": "2024-12-31T00:00:00.000Z"
                      },
                      "organization": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "테스트 학교"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "프로필 권한 목록 조회",
        "tags": [
          "permissions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Thinking API Server",
    "description": "수업 관리 시스템 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT 토큰을 입력하세요",
        "in": "header"
      }
    },
    "schemas": {
...
      "GrantPermissionPayload": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "number",
            "description": "조직 ID",
            "example": 1
          },
          "profileId": {
            "type": "number",
            "description": "권한을 부여할 프로필 ID",
            "example": 1
          },
          "permission": {
            "type": "string",
            "description": "부여할 권한",
            "enum": [
              "course:create",
              "course:update",
              "course:delete",
              "course:enrollment:manage",
              "course:attendance:manage",
              "course:instructor:assign",
              "course:view",
              "course:session:manage",
              "course:class:manage",
              "permission:manage"
            ],
            "example": "course:create"
          },
          "expiresAt": {
            "type": "string",
            "description": "권한 만료일 (YYYY-MM-DD 형식)",
            "example": "2024-12-31"
          }
        },
        "required": [
          "organizationId",
          "profileId",
          "permission"
        ]
      },
      "UpdatePermissionPayload": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string",
            "description": "권한 만료일 (YYYY-MM-DD 형식)",
            "example": "2024-12-31"
          }
        }
      },
...
    }
  }
}